fixed gain computer mix up
